pyproject.toml
[project]
name = "ariel"
version = "2.0.0"
description = "A tool for dubbing videos."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.118.0",
    "google-cloud-aiplatform>=1.119.0",
    "google-cloud-storage>=2.19.0",
    "jinja2>=3.1.4",
    "google-genai>=0.5.4",
    "moviepy>=2.2.1",
    "demucs>=4.0.1",
    "soundfile>=0.13.1",
    "python-multipart>=0.0.20",
]

[dependency-groups]
dev = [
    "basedpyright>=1.31.6",
    "ruff>=0.13.3",
]

[tool.ruff]
fix = true
line-length = 80
preview = true
indent-width = 2
target-version = "py313"

[tool.ruff.format]
preview = true
docstring-code-format = true

[tool.ruff.lint.pydocstyle]
convention = "google"
